Added: MAL types, MAL XML parser, tasks (base), manager (base).

+++ body/src/IMyAnimeListAPI.java
@@ -6,8 +6,8 @@ import java.io.IOException;
 public interface IMyAnimeListAPI extends Runnable{
     void setProfile(IProfile profile);
     String searchEntry(String name, String type) throws IOException;
-    void addEntry(int id, String data, String type);
-    void updateEntry(int id, String data, String type);
-    void deleteEntry(int id, String data, String type);
+    void addEntry(long id, String data, String type);
+    void updateEntry(long id, String data, String type);
+    void deleteEntry(long id, String data, String type);
     String getList(String type) throws IOException;
 }
